chore: rework sample detail directory & cleanup redux
