+++ src/js/app/websocket/reduxHandler.ts
@@ -3,7 +3,6 @@ import { wsUpdateAnalysis } from "../../analyses/actions";
 import { wsInsertHistory, wsInsertIndex, wsUpdateIndex } from "../../indexes/actions";
 import { wsRemoveOTU, wsUpdateOTU } from "../../otus/actions";
 import { wsInsertReference, wsUpdateReference } from "../../references/actions";
-import { wsInsertSample, wsRemoveSample, wsUpdateSample } from "../../samples/actions";
 
 function actionCreatorWrapper(actionCreator) {
     return (state, message) => actionCreator(message.data);
@@ -13,7 +12,6 @@ const inserters = {
     history: actionCreatorWrapper(wsInsertHistory),
     indexes: actionCreatorWrapper(wsInsertIndex),
     references: actionCreatorWrapper(wsInsertReference),
-    samples: actionCreatorWrapper(wsInsertSample),
 };
 
 const updaters = {
@@ -27,12 +25,10 @@ const updaters = {
     indexes: actionCreatorWrapper(wsUpdateIndex),
     otus: actionCreatorWrapper(wsUpdateOTU),
     references: actionCreatorWrapper(wsUpdateReference),
-    samples: actionCreatorWrapper(wsUpdateSample),
 };
 
 const removers = {
     otus: actionCreatorWrapper(wsRemoveOTU),
-    samples: actionCreatorWrapper(wsRemoveSample),
 };
 
 const modifiers = {
