+++ src/js/samples/utils.ts
@@ -1,0 +1,17 @@
+import { LibraryType } from "@samples/types";
+
+export const workflowStates = {
+    NONE: "none",
+    PENDING: "pending",
+    READY: "ready",
+};
+
+const libraryTypes = {
+    normal: "Normal",
+    srna: "sRNA",
+    amplicon: "Amplicon",
+};
+
+export function getLibraryTypeDisplayName(libraryType: LibraryType) {
+    return libraryTypes[libraryType];
+}
